🚚 chore: 번역 관련 컴포넌트 정리 및 리네이밍
