+++ src/app/post/_components/MapComponent.tsx
@@ -1,62 +1,39 @@
-import MapComponent from '@/app/post/_components/MapComponent2';
-import Image from 'next/image';
-import { parsedData } from '../_types/parsedData';
+'use client';
 
-export const ContentRenderer = ({
-  parsedArray,
-}: {
-  parsedArray: parsedData[];
-}) => {
+import { GoogleMap, Marker, useLoadScript } from '@react-google-maps/api';
+import React from 'react';
+
+type mapData = {
+  lat: string | undefined;
+  lng: string | undefined;
+  name: string | undefined;
+  address: string | undefined;
+};
+
+const MapComponent = ({ lat, lng, name, address }: mapData) => {
+  const { isLoaded, loadError } = useLoadScript({
+    googleMapsApiKey: process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY!,
+  });
+  if (!isLoaded) return <div>Loading Google Maps...</div>;
+  if (loadError) return <div>Error loading Google Maps</div>;
+
   return (
-    <div className="grid gap-4">
-      {parsedArray.map((item, index) => {
-        // `data-type="map"`인지 확인
-        if (
-          item.type === 'element' &&
-          item.attributes &&
-          item.attributes['data-type'] === 'map'
-        ) {
-          const { lat, lng, name, address } = item.attributes;
-          return (
-            <MapComponent
-              key={index}
-              lat={lat}
-              lng={lng}
-              name={name}
-              address={address}
-            />
-          );
-        }
-
-        // 일반 텍스트나 이미지 렌더링
-        if (
-          item.type === 'element' &&
-          item.tag === 'img' &&
-          item.attributes &&
-          item.attributes.src
-        ) {
-          return (
-            <Image
-              key={index}
-              src={item.attributes.src}
-              alt="Content"
-              width={0}
-              height={0}
-              sizes="100vw"
-              style={{ width: '100%', height: 'auto' }}
-            />
-          );
-        }
-
-        if (item.type === 'element' && item.tag === 'p') {
-          return (
-            <p key={index} dangerouslySetInnerHTML={{ __html: item.content }} />
-          );
-        }
-
-        // 기본적으로 아무것도 렌더링하지 않음
-        return null;
-      })}
+    <div className="">
+      <GoogleMap
+        center={{ lat: parseFloat(lat!), lng: parseFloat(lng!) }}
+        zoom={14}
+        options={{
+          disableDefaultUI: true,
+        }}
+        mapContainerStyle={{ width: '100%', height: '217px', borderRadius: 8 }}
+      >
+        <Marker
+          position={{ lat: parseFloat(lat!), lng: parseFloat(lng!) }}
+          title={name}
+        />
+      </GoogleMap>
     </div>
   );
 };
+
+export default MapComponent;
